Plugin Refactor - complete UI refactor Overhaul! Looking and working great. Next - Landscape seal polish and shipping for testing.

+++ Plugins/Digger/Source/Digger/Public/VoxelChunk.h
@@ -196,7 +196,7 @@ private:
 private:
     int16 TerrainGridSize; // Default size of 1 meter
     int8 Subdivisions;
-    int16 VoxelSize;
+    float VoxelSize;
     int16 SectionIndex;
 
 private:
